Era must not be NaN

+++ src/shrubbery/data/ingest.py
@@ -72,7 +72,8 @@ def read_parquet_and_unpack(
     data[feature_cols] = (
         data[feature_cols].apply(lambda x: x / 4.0).astype(np.float32)
     )
-    current_era = napi.get_current_round()  # not quite the "current era"
+    # Not quite the "current era", but guaranteed to be ahead
+    current_era = np.float32(napi.get_current_round())
     data_column_era = data[COLUMN_ERA]
     data_column_era = np.where(
         data_column_era == 'X', current_era, data_column_era
